+++ Back_End/src/main/java/com/adaptivelearning/server/Repository/AnswerRepository.java
@@ -3,6 +3,6 @@ package com.adaptivelearning.server.Repository;
 import com.adaptivelearning.server.Model.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer,Integer> {
-    Answer findByAnswerId(Integer answerId);
+public interface AnswerRepository extends JpaRepository<Answer,Long> {
+    Answer findByAnswerId(Long answerId);
 }
