+++ Back_End/src/main/java/com/adaptivelearning/server/Repository/QuestionRepository.java
@@ -3,6 +3,6 @@ package com.adaptivelearning.server.Repository;
 import com.adaptivelearning.server.Model.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question,Integer> {
-    Question findByQuestionId(Integer questionId);
+public interface QuestionRepository extends JpaRepository<Question,Long> {
+    Question findByQuestionId(Long questionId);
 }
