+++ Back_End/src/main/java/com/adaptivelearning/server/Repository/QuizRepository.java
@@ -3,6 +3,6 @@ package com.adaptivelearning.server.Repository;
 import com.adaptivelearning.server.Model.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz,Integer> {
-    Quiz findByQuizId(Integer quizId);
+public interface QuizRepository extends JpaRepository<Quiz,Long> {
+    Quiz findByQuizId(Long quizId);
 }
