+++ Back_End/src/main/java/com/adaptivelearning/server/Repository/SectionRepository.java
@@ -3,9 +3,9 @@ package com.adaptivelearning.server.Repository;
 import com.adaptivelearning.server.Model.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section, Integer>{
+public interface SectionRepository extends JpaRepository<Section, Long>{
 	
-	Section findBySectionId(Integer sectionId);
+	Section findBySectionId(Long sectionId);
 	
 	Section findByTitle(String title);
