+++ Back_End/src/main/java/com/adaptivelearning/server/Repository/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUserId(Integer userId);
+    User findByUserId(Long userId);
 
     User findByEmail(String email);
